# perf: Add fast path for `printNumber`