--- a/src/utils/print-number.js
+++ b/src/utils/print-number.js
@@ -1,9 +1,6 @@
 function printNumber(rawNumber) {
   if (rawNumber.length === 1) {
-    const code = rawNumber.charCodeAt(0);
-    if (code >= /* "1" */ 48 && code <= /* "9" */ 57) {
-      return rawNumber;
-    }
+    return rawNumber;
   }
   return (
     rawNumber