## Description

The `printNumber` function has become complex for edge cases. From my experience, most of the numeric literals that appear in the source code are single-digit numbers. Therefore, it makes sense to provide a fast path for those cases.

Below are the total execution times of the printNumber function when running on the `main` branch and this PR for the source code of Prettier, Babel, and Excalidraw:

|repo |  `main`  | this PR |
|--------|-------|---|
|prettier | 1.9389689999988775 ms   | 0.759684000002693 ms |
|babel | 5.473972999931675 ms | 1.4176010000458064 ms |
|excalidraw | 8.948650999990122 ms | 4.885992999990378 ms |


<details>
    <summary>Numbers of number literals in the source code(we can see that there are many single digit numbers):</summary>
    
**Prettier:**

```
Map(22) {
  '0' => 656,
  '1' => 533,
  '2' => 135,
  '2097152' => 1,
  '1000' => 2,
  '80' => 2,
  '3' => 18,
  '25' => 1,
  '4' => 10,
  '14' => 1,
  '7' => 2,
  '5' => 5,
  '0.25' => 1,
  '6' => 1,
  '8' => 1,
  '19' => 1,
  '20' => 1,
  '0x1f' => 1,
  '0x7f' => 1,
  '0x9f' => 1,
  '0x300' => 1,
  '0x36f' => 1
}
```

**Babel:**

```
Map(94) {
  '0' => 1519,
  '1' => 1316,
  '2021' => 1,
  '3' => 165,
  '4' => 94,
  '2' => 325,
  '12' => 6,
  '18' => 1,
  '9' => 13,
  '14' => 3,
  '17' => 6,
  '20' => 4,
  '21' => 1,
  '22' => 2,
  '23' => 1,
  '5' => 55,
  '16' => 20,
  '6' => 45,
  '1e3' => 1,
  '1e6' => 1,
  '100' => 4,
  '50' => 6,
  '10' => 53,
  '8' => 36,
  '7' => 125,
  '25' => 5,
  '30' => 1,
  '1000' => 2,
  '0.5' => 1,
  '200' => 2,
  '71' => 1,
  '10.1' => 1,
  '64' => 2,
  '4096' => 1,
  '500_000' => 1,
  '60702' => 1,
  '123' => 3,
  '49' => 4,
  '59' => 1,
  '11' => 8,
  '70' => 1,
  '45' => 1,
  '6.9' => 1,
  '55' => 1,
  '1.2' => 3,
  '255' => 1,
  '0x10ffff' => 1,
  '15' => 2,
  '0x7f' => 2,
  '0xffff' => 5,
  '0x10000' => 4,
  '0xaa' => 2,
  '0x3ff' => 4,
  '0xfc00' => 5,
  '0xdc00' => 2,
  '0xd800' => 3,
  '0xeac7' => 1,
  '0b0000' => 3,
  '0b0001' => 3,
  '0b0010' => 3,
  '0b0100' => 3,
  '0b1000' => 3,
  '0x0009' => 1,
  '0x000b' => 1,
  '0x000c' => 1,
  '0x2000' => 1,
  '0x2001' => 1,
  '0x2002' => 1,
  '0x2003' => 1,
  '0x2004' => 1,
  '0x2005' => 1,
  '0x2006' => 1,
  '0x2007' => 1,
  '0x2008' => 1,
  '0x2009' => 1,
  '0x200a' => 1,
  '0x202f' => 1,
  '0x205f' => 1,
  '0x3000' => 1,
  '0xfeff' => 1,
  '2.1' => 1,
  '4.5' => 2,
  '3.1' => 1,
  '5000' => 1,
  '0o666' => 1,
  '0o266' => 1,
  '0o466' => 1,
  '42' => 7,
  '60' => 3,
  '90' => 1,
  '113' => 1,
  '1.1' => 1,
  '28' => 2,
  '32' => 2
}
```

**excalidraw:**

```
Map(896) {
  '3000' => 4,
  '10' => 468,
  '4' => 236,
  '2' => 1183,
  '0' => 3172,
  '100' => 757,
  '200' => 136,
  '95' => 4,
  '0.1' => 18,
  '1' => 2098,
  '0.3' => 15,
  '93' => 3,
  '180' => 15,
  '40' => 99,
  '5' => 150,
  '1699' => 1,
  '1813275999' => 1,
  '209.72304760646858' => 2,
  '338.83587294718825' => 2,
  '116.42036295658873' => 2,
  '103.65107323746608' => 2,
  '1445523839' => 1,
  '92.28090097254909' => 4,
  '7.105427357601002e-15' => 4,
  '154.72281841151394' => 2,
  '19.199290805487394' => 2,
  '155.45758928571422' => 2,
  '79.43840749607878' => 2,
  '99.89923520113778' => 2,
  '103.6510732374661' => 2,
  '40.317783799181804' => 2,
  '79.1587107641305' => 2,
  '39.037226329125524' => 2,
  '21.285677238400705' => 2,
  '3901' => 1,
  '540959103' => 1,
  '249.48446738689245' => 2,
  '374.851387389359' => 2,
  '88.21658171083376' => 2,
  '113.8575037534261' => 2,
  '1513238033' => 1,
  '0.22814350714115691' => 4,
  '43.414939319563715' => 4,
  '0.06274947619197979' => 2,
  '42.63794490105306' => 2,
  '0.21453039840335475' => 2,
  '52.43469208825097' => 2,
  '4.315205554872581' => 2,
  '56.66774540453215' => 2,
  '20.089784992984285' => 2,
  '60.25027917349701' => 2,
  '46.7532926683984' => 2,
  '61.365826671969444' => 2,
  '72.22851104292477' => 2,
  '59.584691681394986' => 2,
  '85.76368213524371' => 2,
  '55.325139565662596' => 2,
  '87.67263486434864' => 2,
  '51.7342924478499' => 2,
  '87.94074036468018' => 2,
  '43.84700272879395' => 2,
  '87.73030872197806' => 2,
  '36.195582644606276' => 2,
  '87.2559282533682' => 2,
  '43.758132174307036' => 2,
  '81.5915337527493' => 2,
  '47.984890854524416' => 2,
  '69.66352776578219' => 2,
  '50.4328058257654' => 2,
  '42.481213744224995' => 2,
  '52.49167708145666' => 4,
  '20.68789182864576' => 2,
  '51.26396751574663' => 2,
  '3.5475921483286084' => 2,
  '47.099726468136254' => 2,
  '0.2758413461535838' => 4,
  '43.46664538034193' => 2,
  '1635' => 1,
  '1383184881' => 1,
  '249.02524930453623' => 2,
  '398.8804363713438' => 2,
  '88.30808627974527' => 8,
  '9.797916664247975' => 4,
  '683951089' => 2,
  '2.1538602707609424' => 4,
  '2.326538897826852' => 4,
  '1.751753055375216' => 4,
  '12.359939318521995' => 4,
  '5.028526743934819' => 4,
  '25.710950037209347' => 4,
  '7.012921076245119' => 4,
  '46.6269757640547' => 4,
  '7.193749997581346' => 4,
  '71.03526003420632' => 4,
  '5.930375670950649' => 4,
  '85.2899738827162' => 4,
  '1.3342483900732343' => 4,
  '2.6041666666666288' => 10,
  '1722' => 1,
  '303290783' => 1,
  ... 796 more items
}
```

</details>




## Checklist

<!-- Please ensure you’ve done all of these things (if applicable). -->
<!-- You can replace the `[ ]` with `[x]` to mark each task as done. -->

- [x] I’ve read the [contributing guidelines](https://github.com/prettier/prettier/blob/main/CONTRIBUTING.md).

<!-- Please DO NOT remove the playground link -->

**✨[Try the playground for this PR](https://prettier.io/playground-redirect)✨**